+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 16. 【server】节点分发支持 file 类型项目
 17. 【agent】项目新增配置日志文件输出到指定目录
 18. 【server】构建产物目录支持通配符`AntPathMatcher`模式 （感谢@saysay [Gitee issues I455FM](https://gitee.com/dromara/Jpom/issues/I455FM) ）
+19. 【server】新增 h2 数据库缓存大小配置 [CACHE_SIZE](http://www.h2database.com/html/features.html#cache_settings) `db.cacheSize`
 
 ### 解决BUG、优化功能